Extract view helpers to reusable module
Code is shared between ItemHTML and PageHTML,
and they really do not belong in web anyway
since it is usable outside web scope as well.
