+++ lib/mse25_web/controllers/item_html.ex
@@ -1,5 +1,7 @@
 defmodule Mse25Web.ItemHTML do
   use Mse25Web, :html
 
+  import Mse25.Utils
+
   embed_templates "item_html/*"
 end
